requirements.txt
# Function dependencies, for example:
# package>=version
python-dateutil